+++ tsqlike/__about__.py
@@ -1,3 +1,3 @@
 """ Version number in a single place """
 
-__version__ = "1.1.3"
+__version__ = "1.1.4"