+++ tsqlike/tsqlike.py
@@ -94,58 +94,69 @@ def close_file(file):
 
 
 # ------------------------------------------------------------------------------------------------ #
-def str_to_type(s, use_none=False):
+def str_to_type(s, convert_bool=True, convert_numbers=True, use_none=False):
     """ Convert string s to a proper type: int, float or boolean """
 
     # Convert '' - empty strings to None?
     if s == '' and use_none:
         return None
 
-    if s in ('True', 'true'):
-        return True
+    if convert_bool:
+        if s in ('True', 'true'):
+            return True
 
-    if s in ('False', 'false'):
-        return False
+        if s in ('False', 'false'):
+            return False
 
-    try:
-        return float(s) if '.' in s or ',' in s else int(s)     # to float and int
-    except (ValueError, TypeError):
-        return s                                                # no conversion possible -> string
+    if convert_numbers:
+        try:
+            return float(s) if '.' in s or ',' in s else int(s) # to float and int
+        except (ValueError, TypeError):
+            return s                                            # no conversion possible -> string
+
+    return s
 
 
 # ------------------------------------------------------------------------------------------------ #
-def read_csv(in_file=None, encoding=None, newline='', name='', detect_str_types=False,
+def read_csv(in_file=None, encoding=None, newline='', name='',
+             convert_bool=True, convert_numbers=True, use_none=False,
              dialect='excel', **fmtparams):
+
     """
     Read CSV from a file and import into a Table object
 
-    :param in_file:             Filename to read CSV from
-    :param encoding:            Character encoding
-    :param newline:             UNIX/Windows/Mac style line ending
-    :param name:                Table name to assign
-    :param detect_str_types:    Detect and correct types of data, default - False
-    :param dialect:             CSV dialect, e.g: excel, unix
-    :**fmtparams:               Various optional CSV parameters:
-        :param delimiter:       CSV field delimiter
-        :param quotechar:       CSV quote character
-        :param quoting:         CSV quote style
+    :param in_file:         Filename to read CSV from
+    :param encoding:        Character encoding
+    :param newline:         UNIX/Windows/Mac style line ending
+    :param name:            Table name to assign
+    :param convert_bool:    Autoconvert strings into Booleans
+    :param convert_numbers: Autoconvert strings to integers or float
+    :param use_none:        Autoconvert empty strings to None type
+    :param dialect:         CSV dialect, e.g: excel, unix
+    :**fmtparams:           Various optional CSV parameters:
+        :param delimiter:   CSV field delimiter
+        :param quotechar:   CSV quote character
+        :param quoting:     CSV quote style
     :return: Table
     """
 
     f = open_file(in_file, file_mode='r', encoding=encoding, newline=newline)
     _data = csv.reader(f, dialect=dialect, **fmtparams)
-    t = Table(data=list(_data), name=name, detect_str_types=detect_str_types)
+    t = Table(data=list(_data), name=name,
+              convert_bool=convert_bool, convert_numbers=convert_numbers, use_none=use_none)
     close_file(f)
     return t
 
 
 # -------------------------------------------------------------------------------------------- #
-def read_json(in_file=None, name='', detect_str_types=False):
+def read_json(in_file=None, name='', convert_bool=True, convert_numbers=True, use_none=False):
     """ Read JSON data from file
 
     :param in_file:         Filename to read JSON from
     :param name:            Table name to assign
-    :param detect_str_types:    Detect and correct types of data, default - False
+    :param convert_bool:    Autoconvert strings into Booleans
+    :param convert_numbers: Autoconvert strings to integers or float
+    :param use_none:        Autoconvert empty strings to None type
     :return                 Table
     """
 
@@ -155,7 +166,8 @@ def read_json(in_file=None, name='', detect_str_types=False):
         _data = json.load(f)
     except (IOError, OSError) as _err:
         print(f'FATAL@Table.read_json(): Unable to load JSON structure: {_err}')
-    t = Table(data=_data, name=name, detect_str_types=detect_str_types)
+    t = Table(data=_data, name=name,
+              convert_bool=convert_bool, convert_numbers=convert_numbers, use_none=use_none)
     close_file(f)
     return t
 
@@ -224,7 +236,9 @@ class Table:
     """
 
     # -------------------------------------------------------------------------------------------- #
-    def __init__(self, data=None, name=None, detect_str_types=False, use_none=False):
+    def __init__(self, data=None, name=None,
+                 convert_bool=True, convert_numbers=True, use_none=False):
+
         self.timestamp = int(time.time())
         self.name = name or str(self.timestamp)
 
@@ -235,13 +249,19 @@ class Table:
             self.cols = 0
         elif isinstance(data, list) and len(data):
             if isinstance(data[0], dict):                   # list(dicts())
-                self.import_list_dicts(data, detect_str_types=detect_str_types, use_none=use_none)
+                self.import_list_dicts(data, convert_bool=convert_bool,
+                                       convert_numbers=convert_numbers, use_none=use_none)
             if isinstance(data[0], list):                   # list(lists())
-                self.import_list_lists(data, detect_str_types=detect_str_types, use_none=use_none)
+                self.import_list_lists(data, convert_bool=convert_bool,
+                                       convert_numbers=convert_numbers, use_none=use_none)
         elif isinstance(data, dict) and len(data):
-            print(type(next(iter(data))))
             if isinstance(data[next(iter(data))], list):    # dict(lists()):
-                self.import_dict_lists(data, detect_str_types=detect_str_types, use_none=use_none)
+                self.import_dict_lists(data, convert_bool=convert_bool,
+                                        convert_numbers=convert_numbers, use_none=use_none)
+        elif isinstance(data, Table):                       # Table()
+            # Import/export is done to allow type conversion to happen
+            self.import_list_lists(data.export_list_lists(), convert_bool=convert_bool,
+                                   convert_numbers=convert_numbers, use_none=use_none)
         else:
             raise ValueError('FATAL@Table.__init__: Unexpected data format')
 
@@ -267,16 +287,19 @@ class Table:
         self.cols = self.rows and len(self.table[0]) or 0
 
     # -- Import methods -------------------------------------------------------------------------- #
-    def import_list_dicts(self, data, name=None, detect_str_types=False, use_none=False):
+    def import_list_dicts(self, data, name=None,
+                          convert_bool=True, convert_numbers=True, use_none=False):
+
         """
         Import a list of dictionaries
 
-        :alias:                     import_thashes()
-        :param data:                Data to import formatted as list of dictionaries
-        :param name:                If not None, set it as the Table name
-        :param detect_str_types:    Detect and correct types of data, default - False
-        :param use_none:            Use None type for empty, i.e. '', strings
-        :return:                    self
+        :alias:                 import_thashes()
+        :param data:            Data to import formatted as list of dictionaries
+        :param name:            If not None, set it as the Table name
+        :param convert_bool:    Autoconvert strings into Booleans
+        :param convert_numbers: Autoconvert strings to integers or float
+        :param use_none:        Autoconvert empty strings to None type
+        :return:                self
         """
 
         # Set a new Table name if requested
@@ -287,8 +310,8 @@ class Table:
             self.header = [self.name + TNAME_COLUMN_DELIMITER + str(f)
                            if TNAME_COLUMN_DELIMITER not in str(f) else f for f in (data[0].keys())]
 
-            self.table = [list(r.values()) for r in data] if not detect_str_types else [
-                [str_to_type(v, use_none) for v in r.values()] for r in data]
+            self.table = [[str_to_type(v, convert_bool, convert_numbers, use_none) for v in r.values()]
+                          for r in data]
 
         else:
             raise ValueError('FATAL@Table.import_list_dicts: Unexpected data format')
@@ -299,7 +322,9 @@ class Table:
         return self
 
     # -------------------------------------------------------------------------------------------- #
-    def import_dict_lists(self, data, name=None, detect_str_types=False, use_none=False):
+    def import_dict_lists(self, data, name=None,
+                          convert_bool=True, convert_numbers=True, use_none=False):
+
         """
         Import a dictionary of lists
         """
@@ -317,7 +342,7 @@ class Table:
 
             for c, f in enumerate(data.keys()):
                 for r, v in enumerate(data[f]):
-                    self.table[r][c] = v if not detect_str_types else str_to_type(v, use_none)
+                    self.table[r][c] = str_to_type(v, convert_bool, convert_numbers, use_none)
             self._redimension()
         else:
             raise ValueError('FATAL@Table.import_dict_lists: Unexpected data format')
@@ -327,14 +352,17 @@ class Table:
 
     # -------------------------------------------------------------------------------------------- #
     def import_list_lists(self, data, header=True, name=None,
-                          detect_str_types=False, use_none=False):
+                          convert_bool=True, convert_numbers=True, use_none=False):
+
         """
         Import list(list_1(), list_n()) with optional first row as the header
 
         :param data:            Data to import formatted as list of lists
         :param header:          If true, data to import HAS a header
         :param name:            If not None, set it as the Table name
-        :param detect_str_types:    Detect and correct types of data, default - False
+        :param convert_bool:    Autoconvert strings into Booleans
+        :param convert_numbers: Autoconvert strings to integers or float
+        :param use_none:        Autoconvert empty strings to None type
         :return:                self
         """
 
@@ -344,10 +372,8 @@ class Table:
 
         if isinstance(data, list) and len(data) and isinstance(data[0], list):
             # TODO: Check all rows to be equal length
-            if not detect_str_types:
-                self.table = data[1:] if header else data
-            else:
-                self.table = [[str_to_type(v, use_none) for v in r] for r in data[1:]]
+            self.table = [[str_to_type(v, convert_bool, convert_numbers, use_none) for v in r]
+                          for r in data[1:]]
 
             self._redimension()