+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+* **2024.09.20    tSQLike-1.1.4**
+  * `detect_str_types` obsolete by `convert_bool=True`, `convert_numbers=True`, `use_none=False`, see `str_to_type()`
+  * `Table()` object, may be passed to a new `Table()` object, i.e., now possible: `Table(Table())`
+
 * **2024.09.19    tSQLike-1.1.3**
   * `detect_types=False` renamed to `detect_str_types=False`
   * `str_to_type()` `use_none=False` whether to convert empty strings to `None` or not; Boolean conversion fix